+++ pages/home.js
@@ -1,18 +1,47 @@
-import React from "react";
-import RoadmapCard from "../components/RoadmapCard";
+import React, { useState, useEffect } from "react";
+import Head from "next/head";
+import Image from "next/image";
+import styles from "../styles/Home.module.css";
+import AppBar from "../components/AppBar";
+import BasicCard from "../components/BasicCard";
 
-function Home() {
-  const image = "https://miro.medium.com/max/960/1*PFEk9y9K_k5Ql58pC567-A.png";
-  return (
-    <div>
-      <h1>Hola Cecy</h1>
-      <RoadmapCard
-        title="Rails"
-        description="best way of learning Rails and React"
-        urlImage={image}
-      />
-    </div>
-  );
+export default function Home() {
+	const [data, setData] = useState();
+
+	useEffect(() => {
+		const getCurriculums = async () => {
+			const response = await fetch("/api/curriculums");
+			const getResponse = await response.json();
+			setData(getResponse);
+		};
+		getCurriculums();
+	}, []);
+
+	return (
+		<div className={styles.container}>
+			<Head>
+				<title> 🔥 Paraffin 🔥 </title>
+				<meta name="description" content="Generated by create next app" />
+			</Head>
+			<AppBar />
+			<main className={styles.main}>
+				<h1 className={styles.title}>🔥 Paraffin 🔥</h1>
+				<br></br>
+				<div>
+					{" "}
+					<h2>
+						{" "}
+						El camino a convertirse en desarrollador nunca fue tan fácil{" "}
+					</h2>
+				</div>
+				<div>
+					{data?.map((item, i) => (
+						<p className={styles.description} key={item.id}>
+							<BasicCard title={item.name} description={item.description} />
+						</p>
+					))}
+				</div>
+			</main>
+		</div>
+	);
 }
-
-export default Home;
