+++ pages/index.js
@@ -5,35 +5,23 @@ import styles from "../styles/Home.module.css";
 import AppBar from "../components/AppBar";
 import BasicCard from "../components/BasicCard";
 
-export default function Home() {
-	const [data, setData] = useState();
+import Home from "./home";
 
-	useEffect(() => {
-		const getCurriculums = async () => {
-			const response = await fetch("/api/curriculums");
-			const getResponse = await response.json();
-			setData(getResponse);
-		};
-		getCurriculums();
-	}, []);
+export default function Index() {
+	// const [data, setData] = useState();
+
+	// useEffect(() => {
+	// 	const getCurriculums = async () => {
+	// 		const response = await fetch("/api/curriculums");
+	// 		const getResponse = await response.json();
+	// 		setData(getResponse);
+	// 	};
+	// 	getCurriculums();
+	// }, []);
 
 	return (
 		<div className={styles.container}>
-			<Head>
-				<title> 🔥 Paraffin 🔥 </title>
-				<meta name="description" content="Generated by create next app" />
-			</Head>
-			<AppBar />
-			<main className={styles.main}>
-				<h1 className={styles.title}>🔥 Paraffin 🔥</h1>
-
-				{data?.map((item, i) => (
-					<p className={styles.description} key={item.id}>
-						{/* {item.name} */}
-						<BasicCard title={item.name} description={item.description} />
-					</p>
-				))}
-			</main>
+			<Home />
 		</div>
 	);
 }
